requirements.txt
absl-py==0.11.0
alembic==1.4.3
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.0
astunparse==1.6.3
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.7.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudinary==1.23.0
colorama==0.4.3
command-not-found==0.3
config==0.5.0.post0
confusable-homoglyphs==3.2.0
cryptography==3.2.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.3
django-admin==2.0.1
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-contrib-comments==1.9.2
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-generic-aggregation==0.4.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-on-heroku==1.1.2
django-registration==3.1.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-simple-ratings==0.3.3
django-six==1.0.4
django-star-ratings==0.9.1
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==2.0.1
netifaces==0.10.4
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==8.0.1
pipenv==2020.8.13
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
screen==1.0.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
swapper==1.1.2.post1
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
testresources==2.0.1
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xkit==0.0.0
xlwt==1.3.0
zipp==3.4.0
